controler class is added
